+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - __DEPRECATION:__ Force append `/auth` in server url.
 
 ### Fixed
+- required action providerId and alias can be different now
+- providerId of required actions can be updated now
 
 ## [2.4.0] - 2020-10-05